fix: workaround for getBlobIdFromAttributeFilePath
BlobStoreSupport#getBlobIdFromAttributeFilePath is new in 3.15, and there is a small issue in how we want to use it here.

Previously, getBlobIdStream excluded temp blobs from the result stream. This behavior is changed in 3.15 - getBlobIdStream should include temp blobs. getBlobIdFromAttributeFilePath does not correctly return the BlobId for a temporary blob that already exists, it's only intended to be used for new blobs.