Update to support NXRM 3.15 (#25)
* Work for 3.15 compatibility, including refactoring to take advantage of upstream support classes.

* travis config now uses openjdk8

* fix: workaround for getBlobIdFromAttributeFilePath

BlobStoreSupport#getBlobIdFromAttributeFilePath is new in 3.15, and there is a small issue in how we want to use it here.

Previously, getBlobIdStream excluded temp blobs from the result stream. This behavior is changed in 3.15 - getBlobIdStream should include temp blobs. getBlobIdFromAttributeFilePath does not correctly return the BlobId for a temporary blob that already exists, it's only intended to be used for new blobs.

* fix: blobstore must be provided to metrics store prior to start

BlobStoreMetricsStoreSupport#start has a subtle requirement that BlobStoreMetricsSupport#setBlobStore be called first with a non-null argument. The reference to the blobstore is used by a periodic job that checks for soft quota issues, this job is instantiated/scheduled by doStart.