feat: new kar deliverable (#34)
* feat: new kar deliverable

This change set leverages the kar goal from the karaf-maven-plugin to simplify the deliverable. As of this change, we no longer need the edits to nexus core feature files, nor the unorthodox changes to the karaf maven configuration.

The reduction of steps in the Dockerfile is evidence of this simplification; just drop the kar file in the deploy folder is all that we need now.