refactor: establish new key ancestry for deleted blobs
Previously, the DeletedBlobIndex stored the ids of soft-deleted blobs in the root namespace in Google Cloud Datastore.

After this change set, the ids are stored in a new hierarchy that is friendlier for a) multiple google cloud blobstore instances in the same project and b) will support us using Datastore for other blobstore metadata.

Also included is further expansion of the HealthCheck - it can now indicate that a google cloud blobstore has a high number of soft-deleted blobs in need of compaction.