so long circleci badge
Always shows as failed. It appears the badge does not correctly interpret a workflow step that requires input (like ours; after building master, the workflow goes on hold, waiting for one of the maintainers to say "go ahead and release").
If circleci badges can show the status of the last completed step, we could put this back; until then, it has no value to the project.