update google client libraries to latest
Includes some OSGi configuration to adapt to the NXRM environment. A big stumbling block was the guava dependency; the client libraries reference 26+, but only 25 is available in the container.

We also have to explicitly include google-oauth-client as a dependency as it is declared as a mandatory import by google-api-client, but not in the compile/runtime dependency hierarchy.

Fixes #41.