feat: bucket location is now configurable
Previously, the plugin would create the storage bucket using default values for location and storage class from whatever the SDK would use (I believe to be multi-regional, and us).

In this change set, we do the following:

* Allow the region to be passed in via BlobStoreConfiguration
* Present region in the UI to create a BlobStore
* Force a storage class of REGIONAL

The REGIONAL storage class provides the lowest latency and best repository manager performance, given the repository manager instance and the bucket have matching values for region.