on-thread alternative to MultipartUploader
The BufferingUploader follows the same contract as MultipartUploader to break content into fixed size chunks and upload with gzip compression disabled. The storage writes and compose request all occur in serial on the main thread.

One benefit to this option is that there are no calls to Arrays.copyOfRange; every chunk uses the exact same byte[] buffer, so this may have a better memory profile. The runtimes on integration tests are slower than MultipartUpload, so isolating to a separate feature branch at this time.