fix: decrease default multipart upload size to 2 MB (from 5)
After more detailed stress testing, we are finding that a default of 5 MB results in too high of heap usage, particularly for workloads that store very large blobs (docker).
It is better to have the chunk size be too small and hit the compose limit than to have a high value and OOM. As a result, the log message entered when the compose limit is hit was changed to debug and the recommendation to increase was remove.

The compose limit hit count is also now a counter exposed in the system metrics.