refactor: encapsulate datastore-backed daos for better error handling (#66)
* refactor: encapsulate datastore-backed daos for better error handling

This changeset is a pretty significant refactor inspired by #63. GoogleCloudBlobStore#doInit is called when creating a new Blobstore in NXRM. Prior to this, that method simply created the bucket and wrote a small metadata file; if either of those steps would fail due to permissions or something else, that error would be presented in the user interface to the actor. The MetricsStore and DeletedBlobIndex however were managed separately, and no test was done during initialization.

Now, GoogleCloudBlobStore#doInit is changed to attempt to initialize the bucket, the deleted blob index, and the metrics store, individually and in that order. If any fails, a more meaningful exception will bubble up to the actor.