Reduce read ops (#84)
* refactor: prioritize small operations for DeletedBlobIndex

Prior to this changeset, the DeletedBlobIndex used full Entity queries. This is unnecessary, as the BlobIds are the keys themselves.
The getContents() method was a particular offender: it was called frequently to get a count of the number of entries in the index, and the number of entries is potentially unbounded (if no compact blobstore task has executed).
In this change, we switch to key only queries, which count as small operations instead of full document reads.