chore: update for NXRM 3.32
This upgrade ended up requiring a change to the Karaf configuration. Without the bundle configuration changes, the plugin would deploy successfully, but attempts to create a blobstore resulted in Classloader errors.
Karaf was upgraded in NXRM 3.32, and it appears that the grpc-context dependency no longer needs the "Fragment-Host" attribute. After this configuration change, the plugin works as expected.