refactor: migration bucket, location, credential_file attributes
As of NXRM 3.34, the blobstore UI forms expect attributes with specific names; for cloud blobstores, specifically 'bucketName', 'region', and 'credentialFilePath'. In this changeset, a handler is added to the doInit function on the blobstore that detects the presence of attributes using the legacy forms ('bucket', 'location', and 'credential_file', respectively). New Google Cloud Blobstores will be created using the new attribute names.

Co-authored-by: Michael Martz <mmartz@sonatype.com>