fix: increase limit to DeletedBlobIndex#getContents, allow config
DeletedBlobIndex#getContents supplies a limit to the datastore query. In the previous version, this limit was 1000. This method is called by the compact blobstore task.
If the blobstore targeted by the compaction task is significant in size, a limit of 1000 soft-deleted blobs is far too low.

In this change, a new default of 100,000 is supplied as the limit. This was chosen after testing confirmed successful reasonable performance to return. This behavior is now also configurable by providing an alternate value in nexus.properties under the 'nexus.gcs.deletedBlobIndex.contentQueryLimit' property.