bump to 3.57
An upgrade to the enforcer plugin in the parent pom necessitated a few overrides in this project. The plugin has been disabled here as there are a number of recommendations that can't quite be implemented yet. We'll need to restore the enforcer and work through those changes one by one in future efforts.